Move pm-specific numeric equality into its own file
Extract pm_equals_numerical and its helpers value_multisets_match /
values_close from equality/numerical.js into equality/pm-numerical.js.
The new module imports the shared sampling helpers
(prepare_exprs_numerical_eval, replace_numbers_with_parameters,
generate_random_integer, random_cubic) from numerical.js, which now
re-exports the latter two. numerical.js imports pm_equals_numerical
back; the resulting import cycle is safe since every cross-module
reference is a hoisted function declaration used only at call time.
